requirements.txt
ctransformers==0.2.5
sentence-transformers==2.2.2
pinecone-client
langchain
flask
pypdf
python-dotenv
langchain_pinecone
-e .